Microservice_Frontend App/requirements.txt
streamlit
pillow
requests

#Run this file with - pip3 install -r requirements.txt